# docs/swagger.yaml
openapi: 3.0.3
info:
  title: GPTless Actual Intelligence API
  version: 1.0.0
servers:
  - url: https://api.gptless.au
  - url: https://preview.gptless.au
...
security:
  - AccessTokenAuth: []
paths:
...
  /v1/responder/chats:
    get:
      tags: [Responder side]
      summary: list your chats.
      description: lists all the chats belonging to a responder. can be filtered by categories with params

      parameters:
        - $ref: "#/components/parameters/AccessToken"
        - in: query
          name: categories 
          required: false
          style: form
          explode: false
          schema:
            type: array
            items:
              type: string
          description: |
            Filters chats by categories.
            Provide multiple values as a comma-separated list.
      responses:
        "200":
          description: |
            **OK**
            
            Returns an array of chats.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChatSummary"
              example:
                - chatID: 12345
                  title: website design
                  category: development
                  status: open
                  tokens: 30
                  createdAt: "2025-08-14T10:30:00Z"
                - chatID: 54321
                  title: poem request
                  category: writing
                  status: closed
                  tokens: 40
                  createdAt: "2025-08-13T09:00:00Z"

        "401":
          description: |
            **Unauthorized**
            
            Missing or invalid API token.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                UNAUTHORIZED:
                  summary: UNAUTHORIZED
                  value:
                    error: UNAUTHORIZED
                    message: The API token is missing or invalid. If you do not have an API token register for one through the form on our website

        "500":
          description: |
            **Internal Server Error**
            
            There was an internal server error. If this persists contact us at contact@gptless.au.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                INTERNAL_SERVER_ERROR:
                  summary: INTERNAL_SERVER_ERROR
                  value:
                    error: INTERNAL_SERVER_ERROR
                    message: An internal server error occurred.
  /v1/responder/chats/unclaimed:
    get:
      tags: [Responder side]
      summary: list unclaimed chats.
      description: provides a list of unclaimed chats. 

      parameters:
        - $ref: "#/components/parameters/AccessToken"
        - in: query
          name: categories
          required: false
          style: form
          explode: false
          schema:
            type: array
            items:
              type: string
          description: |
            Filters chats by categories.
            Provide multiple values as a comma-separated list.

      responses:
        "200":
          description: |
            **OK**
            
            Returns an array of chats.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChatSummary"
              example:
                - chatID: 12345
                  title: website design
                  category: development
                  status: open
                  tokens: 30
                  createdAt: "2025-08-14T10:30:00Z"
                - chatID: 54321
                  title: poem request
                  category: writing
                  status: closed
                  tokens: 40
                  createdAt: "2025-08-13T09:00:00Z"
        "401":
          description: |
            **Unauthorized**
            
            Missing or invalid API token.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                UNAUTHORIZED:
                  summary: UNAUTHORIZED
                  value:
                    error: UNAUTHORIZED
                    message: The API token is missing or invalid. If you do not have an API token register for one through the form on our website

        "500":
          description: |
            **Internal Server Error**
            
            There was an internal server error. If this persists contact us at contact@gptless.au.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                INTERNAL_SERVER_ERROR:
                  summary: INTERNAL_SERVER_ERROR
                  value:
                    error: INTERNAL_SERVER_ERROR
                    message: An internal server error occurred.
  /v1/responder/chats/{chatID}:
    parameters:
      - $ref: "#/components/parameters/AccessToken"
      - $ref: "#/components/parameters/ChatId"

    get:
      tags: [Responder side]
      summary: get a chat
      description: Given a valid chatID, returns the chat.
      responses:
        "200":
          description: |
            **OK**

            Returns the chat.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Chat"

        "401":
          description: |
            **Unauthorized**
            
            Missing or invalid API token.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                UNAUTHORIZED:
                  summary: UNAUTHORIZED
                  value:
                    error: UNAUTHORIZED
                    message: The API token is missing or invalid. If you do not have an API token register for one through the form on our website

        "403":
          description: |
            **Forbidden**
            
            You do not have access to this chat.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                FORBIDDEN:
                  summary: FORBIDDEN
                  value:
                    error: FORBIDDEN
                    message: You do not have access to this content

        "404":
          description: |
            **Not Found**
            
            ChatID does not exist.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                NOT_FOUND:
                  summary: NOT_FOUND
                  value:
                    error: NOT_FOUND
                    message: The requested resource was not found

        "500":
          description: |
            **Internal Server Error**
            
            There was an internal server error. If this persists contact us at contact@gptless.au.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                INTERNAL_SERVER_ERROR:
                  summary: INTERNAL_SERVER_ERROR
                  value:
                    error: INTERNAL_SERVER_ERROR
                    message: An internal server error occurred.
  /v1/responder/chats/{chatID}/claim:
    post:
      tags: [Responder side]
      summary: claim a chat
      description: |
        Claims an unclaimed chat so the responder can work on it so multiple people don't work on the same job.

      parameters:
        - $ref: "#/components/parameters/AccessToken"
        - $ref: "#/components/parameters/ChatId"

      responses:
        "200":
          description: |
            **OK**
            
            Returns the claimed chat.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Chat"

        "400":
          description: |
            **Bad Request**
            
            Chat cannot be claimed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                BAD_REQUEST:
                  summary: BAD_REQUEST
                  value:
                    error: BAD_REQUEST
                    message: Chat cannot be claimed

        "401":
          description: |
            **Unauthorized**
            
            Missing or invalid API token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                UNAUTHORIZED:
                  summary: UNAUTHORIZED
                  value:
                    error: UNAUTHORIZED
                    message: The API token is missing or invalid. If you do not have an API token register for one through the form on our website

        "403":
          description: |
            **Forbidden**
            
            You do not have access to claim chats.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                FORBIDDEN:
                  summary: FORBIDDEN
                  value:
                    error: FORBIDDEN
                    message: You do not have access to this content

        "404":
          description: |
            **Not Found**
            
            ChatID does not exist.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                NOT_FOUND:
                  summary: NOT_FOUND
                  value:
                    error: NOT_FOUND
                    message: The requested resource was not found

        "409":
          description: |
            **Conflict**
            
            Chat has already been claimed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                CONFLICT:
                  summary: CONFLICT
                  value:
                    error: CONFLICT
                    message: This chat has already been claimed

        "500":
          description: |
            **Internal Server Error**
            
            There was an internal server error. If this persists contact us at contact@gptless.au.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                INTERNAL_SERVER_ERROR:
                  summary: INTERNAL_SERVER_ERROR
                  value:
                    error: INTERNAL_SERVER_ERROR
                    message: An internal server error occurred.
  /v1/responder/chats/{chatID}/messages:
    post:
      tags: [Responder side]
      summary: send a message to the requester 
      description: |
        Sends a responder message in an existing chat. for example asking for clarification or extra details providing the final result.

      parameters:
        - $ref: "#/components/parameters/AccessToken"
        - $ref: "#/components/parameters/ChatId"

      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MessageRequest"

      responses:
        "201":
          description: |
            **Created**
            
            Returns the newly created message.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"

        "400":
          description: |
            **Bad Request**
            
            Missing or invalid message data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                BAD_REQUEST:
                  summary: BAD_REQUEST
                  value:
                    error: BAD_REQUEST
                    message: "Missing or invalid message data: missing field: message"

        "401":
          description: |
            **Unauthorized**
            
            Missing or invalid API token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                UNAUTHORIZED:
                  summary: UNAUTHORIZED
                  value:
                    error: UNAUTHORIZED
                    message: The API token is missing or invalid. If you do not have an API token register for one through the form on our website

        "403":
          description: |
            **Forbidden**
            
            You do not have access to this chat.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                FORBIDDEN:
                  summary: FORBIDDEN
                  value:
                    error: FORBIDDEN
                    message: You do not have access to this content

        "404":
          description: |
            **Not Found**
            
            ChatID does not exist.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                NOT_FOUND:
                  summary: NOT_FOUND
                  value:
                    error: NOT_FOUND
                    message: The requested resource was not found

        "500":
          description: |
            **Internal Server Error**
            
            There was an internal server error. If this persists contact us at contact@gptless.au.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                INTERNAL_SERVER_ERROR:
                  summary: INTERNAL_SERVER_ERROR
                  value:
                    error: INTERNAL_SERVER_ERROR
                    message: An internal server error occurred.
  /v1/responder/chats/{chatID}/fulfill-request:
    post:
      tags: [Responder side]
      summary: send a response with the completed request
      description: |
        Submits the completed work for the current request in a chat.

      parameters:
        - $ref: "#/components/parameters/AccessToken"
        - $ref: "#/components/parameters/ChatId"

      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FulfillRequestRequest"

      responses:
        "201":
          description: |
            **Created**
            
            Returns the submitted fulfillment.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Fulfillment"

        "400":
          description: |
            **Bad Request**
            
            Missing or invalid fulfillment data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                BAD_REQUEST:
                  summary: BAD_REQUEST
                  value:
                    error: BAD_REQUEST
                    message: "Missing or invalid fulfillment data: missing field: responseText"

        "401":
          description: |
            **Unauthorized**
            
            Missing or invalid API token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                UNAUTHORIZED:
                  summary: UNAUTHORIZED
                  value:
                    error: UNAUTHORIZED
                    message: The API token is missing or invalid. If you do not have an API token register for one through the form on our website

        "403":
          description: |
            **Forbidden**
            
            You do not have access to this chat.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                FORBIDDEN:
                  summary: FORBIDDEN
                  value:
                    error: FORBIDDEN
                    message: You do not have access to this content

        "404":
          description: |
            **Not Found**
            
            ChatID does not exist.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                NOT_FOUND:
                  summary: NOT_FOUND
                  value:
                    error: NOT_FOUND
                    message: The requested resource was not found

        "409":
          description: |
            **Conflict**
            
            There is no active request to fulfill.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                CONFLICT:
                  summary: CONFLICT
                  value:
                    error: CONFLICT
                    message: There is no active request to fulfill

        "500":
          description: |
            **Internal Server Error**
            
            There was an internal server error. If this persists contact us at contact@gptless.au.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                INTERNAL_SERVER_ERROR:
                  summary: INTERNAL_SERVER_ERROR
                  value:
                    error: INTERNAL_SERVER_ERROR
                    message: An internal server error occurred.
...
components:
  securitySchemes:
    AccessTokenAuth:
      type: apiKey
      in: header
      name: AccessToken
  parameters:
    ChatId:
      name: chatID
      in: path
      required: true
      schema:
        type: string
      description: chatID
    AccessToken:
      name: AccessToken
      in: header
      required: true
      schema:
        type: string
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9
      description: AccessToken
  schemas:
...
    MessageRequest:
      type: object
      properties:
        message:
          type: string
          example: could you resize the title please.
      required:
        - message
...
    FulfillRequestRequest:
      type: object
      properties:
        responseText:
          type: string
          example: here is the link to the completed code.
        attachments:
          type: array
          items:
            type: string
            example: https://randomDropboxLink.com
      required:
        - responseText
    ChatSummary:
      type: object
      properties:
        chatID:
          type: string
          example: 12345
        title:
          type: string
          example: poem request
        category:
          type: string
          example: writing
        status:
          type: string
          example: open
        Tokens:
          type: number
          example: 30
        createdAt:
          type: string
          format: date-time
          example: "2025-08-14T10:30:00Z"
    Message:
      type: object
      properties:
        messageID:
          type: string
          example: msg_001
        senderType:
          type: string
          example: requester
        message:
          type: string
          example: Please make the tone more professional.
        createdAt:
          type: string
          format: date-time
          example: "2025-08-14T10:35:00Z"
    Request:
      type: object
      properties:
        requestID:
          type: string
          example: req_001
        requestText:
          type: string
          example: Now make a social media version using the same context.
        status:
          type: string
          example: pending
        tokensSpent:
          type: number
          example: 15
        createdAt:
          type: string
          format: date-time
          example: "2025-08-14T10:40:00Z"
    Fulfillment:
      type: object
      properties:
        fulfillmentID:
          type: string
          example: full_001
        requestID:
          type: string
          example: req_001
        responseText:
          type: string
          example: Here is the completed work for your request.
        createdAt:
          type: string
          format: date-time
          example: "2025-08-14T11:10:00Z"
    Chat:
      type: object
      properties:
        chatID:
          type: string
          example: 12345
        requesterUsername:
          type: string
          example: Guy
        responderUsername:
          type: string
          nullable: true
          example: John
        status:
          type: string
          example: open
        createdAt:
          type: string
          format: date-time
          example: "2025-08-14T10:30:00Z"
        messages:
          type: array
          items:
            $ref: "#/components/schemas/Message"
        requests:
          type: array
          items:
            $ref: "#/components/schemas/Request"
...
    Error:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
        message:
          type: string